training dataframe check and selection
